Today I have worked on the easybutor application.
Done some research on the notification and implement as well on notification module.
 now we can send notifications through the application.
currently doing some research to store data on a common node of database and can be used by both the user ie., buyer and seller. for both operations like tracking order by buyer and approval of the order by the seller.